feat!: support mapping projected types (AIRBNB)
This PR extends object mapping to allow mapping a value for a "projected type", where the projection is described by a selection set. This allows mapping aliased values, or (if the selection set and schema are known), responses that were returned by a server.

Mapping with a selection set has a couple of properties that make it different from mapping with just a type:
- while a type can be infinitely recursive, a selection set may not be
- while a type will always be mapped the same wherever it appears, a type projected with a selection set may not be unless it contains the exact same subselections

The result is that Conv factories that can handle selection sets will disable memoization if a selection set is present.

Another property is that GRT's for objects may be keyed differently depending on how they were obtained.
For example, a GRT that was obtained through a required selection set or a `ctx.query` call will be keyed by selection name. This means that an aliased selection must be fetched using the selection name, rather than the field, like `ctx.objectValue.getFoo("aliasedFoo")`

Compare this to a GRT that is created inside a resolver, which at the time that it is created will be keyed by field names (which the engine will then re-map with selection names). This means that even if it the object is being created to satisfy a selection like `{ aliasedFoo: foo }`, the value is not available under "aliasedFoo" and must be fetched using its field name, like `obj.getFoo()`.

This difference in how object GRTs are keyed makes it unclear how they should be mapped. This PR introduces a concept of a `KeyMapping`, which can be supplied to GRTConv to control this behavior. As a practical matter, we can usually infer the right KeyMapping based on the presence of a selection set, though this mapping may be overridden by callers.

Github-Change-Id: 976452
GitOrigin-RevId: 2ee7dc145b7e6ff82b3f7246beff9f1a68f9bd7d